refactor: move constants out of setup files for tests
